extract Dataset model into its own package
